Fix bug "Unable to see images on Home and Creditos page #4" and renamed files to lower case.
