+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:aplicativo_nutricao/data/database_helper.dart';
-import 'package:aplicativo_nutricao/view/LoginPage.dart';
+import 'package:aplicativo_nutricao/view/login_page.dart';
 
 void main() async{
   WidgetsFlutterBinding.ensureInitialized();
