+++ pubspec.yaml
@@ -72,7 +72,7 @@ flutter:
     - assets/images/Denner.png
     - assets/images/Hercules.png
     - assets/images/jeff.png
-    - assets/images/Rafa.png
+    - assets/images/raphael.jpeg
